Fixed encoding nonsense.
The encoding problem was fixed by using BeautifulSoup. This module takes care of
properly encoding data as well as reducing the amount of code required to
retrieve the title of a page.

For this to work you'll have to have BeautifulSoup installed:

    $ pip install beautifulsoup

I also made some changes to the method get_html_and_response(). Instead of
returning some vague object (addinfourl or something) it returns the raw
response body in the second index.

See #11 for more information.

Signed-off-by: Yorick Peterse <yorick@isset.nl>